refactor(properties): own the axum inbound router (hex step 4d)
The properties HTTP API now lives in properties::inbound::axum_router
behind a new 'inbound' feature, following the channels/soup pattern.
properties_service shrinks to a composition shim (concrete type
aliases, auth middleware wiring, OpenAPI doc, openapi bin).

Domain:
- PermissionService port gains check_entity_view_permission (owner
  fast-path, deleted-entity guard, any-access-level check); the impl's
  access-level machinery incl. the thread ownership fallback is now a
  shared helper feeding both view and edit checks
- PropertiesService gains check_entity_view_permission and
  check_entity_edit_permission, replacing properties_service's
  permissions.rs (which reached into email_db_client and raw pools)

Inbound:
- generic PropertiesRouterState<S, A> (properties service + entity
  access service) with FromRef for the team extractor
- all 16 handlers moved and genericized over the service traits; the
  ensure_user_exists auth middleware is injected by the composition
  root as a tower Layer so macro_middleware stays out of the crate
- properties_err_status moves in as the single error->status mapping

properties_service:
- deleted api/properties/** and api/permissions.rs; context.rs is two
  type aliases; swagger ApiDoc now references the crate's handlers
  cross-crate; dependency list pruned from 30 to 9 crates

document_storage_service: constructs PropertiesRouterState::new(...)
(no more raw db pool handed to the properties HTTP layer).
