+++ .github/workspace-dep-closures.json
@@ -3674,8 +3674,6 @@
       "cowlike",
       "decode_jwt",
       "document_sub_type",
-      "email_db_client",
-      "email_utils",
       "entity_access",
       "entity_access_db_utils",
       "invite_email",
@@ -3698,7 +3696,6 @@
       "model_user",
       "models_bulk_upload",
       "models_dcs",
-      "models_email",
       "models_entity_access_management",
       "models_opensearch",
       "models_pagination",
