+++ rust/cloud-storage/properties/src/inbound.rs
@@ -1,1 +1,4 @@
+#[cfg(feature = "inbound")]
+pub mod axum_router;
+#[cfg(feature = "ai_tools")]
 pub mod toolset;
