+++ rust/cloud-storage/properties_service/Cargo.toml
@@ -13,39 +13,15 @@ name = "properties_service_openapi"
 path = "src/openapi.rs"
 
 [dependencies]
-anyhow = { workspace = true }
-async-trait = { workspace = true }
 axum = { workspace = true, features = ["tower-log", "ws"] }
-chrono = { workspace = true }
-email_db_client = { path = "../email_db_client" }
 entity_access = { path = "../entity_access" }
-http-body-util = { workspace = true }
-macro_user_id = { path = "../macro_user_id" }
-model-entity = { path = "../model-entity" }
-macro_auth = { path = "../macro_auth" }
-macro_env = { path = "../macro_env" }
 macro_middleware = { path = "../macro_middleware", default-features = false, features = [
   "auth",
   "cloud_storage",
 ] }
-last_online_tracker = { path = "../last_online_tracker" }
-notification = { path = "../notification" }
-aws-sdk-sqs = { workspace = true }
-macro_uuid = { path = "../macro_uuid" }
-model = { path = "../model" }
-models_permissions = { path = "../models_permissions" }
 models_properties = { path = "../models_properties" }
+notification = { path = "../notification" }
 properties = { path = "../properties" }
-system_properties = { path = "../system_properties" }
-serde = { workspace = true }
 serde_json = { workspace = true }
-serde_with = { workspace = true }
-sqlx = { workspace = true }
-thiserror = { workspace = true }
-tokio = { workspace = true }
-tower = { workspace = true }
-tower-http = { workspace = true }
-tracing = { workspace = true }
 utoipa = { workspace = true }
-uuid = { workspace = true }
 workspace-hack = { version = "0.1", path = "../workspace-hack" }
