--- rust/cloud-storage/properties_service/src/api/properties/definitions/mod.rs
@@ -1,6 +1,0 @@
-pub mod create;
-pub mod delete;
-pub mod list;
-
-// Re-export commonly used types from the list module
-pub use list::{ListPropertiesQuery, PropertyDefinitionResponse};
