--- rust/cloud-storage/properties_service/src/api/properties/entities/get.rs
@@ -1,137 +1,0 @@
-use axum::{
-    Json,
-    extract::{Extension, Path, Query, State},
-    http::StatusCode,
-    response::{IntoResponse, Response},
-};
-use thiserror::Error;
-
-use crate::api::{
-    context::PropertiesHandlerState,
-    properties::entities::types::{
-        EntityPropertiesResponse, EntityQueryParams, retain_caller_visible_tags,
-    },
-    properties::properties_err_status,
-};
-use model::user::UserContext;
-use models_properties::EntityType;
-use properties::{PropertiesErr, PropertiesService};
-
-#[derive(Debug, Error)]
-pub enum GetEntityPropertiesErr {
-    #[error(transparent)]
-    Properties(#[from] PropertiesErr),
-
-    #[error("Entity not found")]
-    MetadataNotFound,
-
-    #[error("{0}")]
-    Permission(#[from] crate::api::permissions::PermissionError),
-}
-
-impl IntoResponse for GetEntityPropertiesErr {
-    fn into_response(self) -> Response {
-        let status_code = match &self {
-            GetEntityPropertiesErr::Properties(e) => properties_err_status(e),
-            // Preserved behavior: a missing entity on the metadata path is a 500.
-            GetEntityPropertiesErr::MetadataNotFound => StatusCode::INTERNAL_SERVER_ERROR,
-            GetEntityPropertiesErr::Permission(e) => e.status_code(),
-        };
-
-        if status_code.is_server_error() {
-            tracing::error!(
-                error = ?self,
-                error_type = "GetEntityPropertiesErr",
-                "Internal server error"
-            );
-        }
-
-        (status_code, self.to_string()).into_response()
-    }
-}
-
-/// Get all properties for an entity
-#[utoipa::path(
-    get,
-    path = "/properties/entities/{entity_type}/{entity_id}",
-    params(
-        ("entity_type" = EntityType, Path, description = "Entity type (user, document, channel, project, thread)"),
-        ("entity_id" = String, Path, description = "Entity ID"),
-        ("include_metadata" = Option<bool>, Query, description = "Whether to include property metadata (default: false)")
-    ),
-    responses(
-        (status = 200, description = "Entity properties retrieved successfully", body = EntityPropertiesResponse),
-        (status = 400, description = "Invalid entity type"),
-        (status = 403, description = "Forbidden - User does not have permission to view this entity"),
-        (status = 404, description = "Entity not found"),
-        (status = 500, description = "Internal server error")
-    ),
-    tag = "Properties"
-)]
-#[tracing::instrument(skip(state, user_context), fields(user_id = %user_context.user_id, entity_type = ?entity_type, include_metadata = query.include_metadata), err)]
-pub async fn get_entity_properties(
-    Path((entity_type, entity_id)): Path<(EntityType, String)>,
-    Query(query): Query<EntityQueryParams>,
-    State(state): State<PropertiesHandlerState>,
-    Extension(user_context): Extension<UserContext>,
-) -> Result<Json<EntityPropertiesResponse>, GetEntityPropertiesErr> {
-    tracing::info!(
-        entity_id = %entity_id,
-        "retrieving entity properties"
-    );
-
-    let entity_ref = models_properties::EntityReference::new(entity_id.clone(), entity_type);
-
-    // Note: This can fail if the entity is marked with "deletedAt"
-    crate::api::permissions::check_entity_view_permission(
-        &state,
-        &user_context.user_id,
-        &entity_ref,
-    )
-    .await?;
-
-    let (user_properties, metadata_properties) = if query.include_metadata {
-        // Fetch user properties and metadata in parallel when metadata is requested
-        let (user_properties_result, metadata_properties_result) = tokio::join!(
-            state
-                .properties_service
-                .get_entity_properties_with_definitions(&entity_id, entity_type),
-            state
-                .properties_service
-                .get_entity_metadata_properties(&entity_id, entity_type)
-        );
-
-        let user_properties = user_properties_result?;
-        let metadata_properties =
-            metadata_properties_result?.ok_or(GetEntityPropertiesErr::MetadataNotFound)?;
-
-        (user_properties, metadata_properties)
-    } else {
-        // Only fetch user properties when metadata not requested - no parallel task needed
-        tracing::debug!("skipping metadata properties due to include_metadata=false");
-        let user_properties = state
-            .properties_service
-            .get_entity_properties_with_definitions(&entity_id, entity_type)
-            .await?;
-
-        (user_properties, vec![])
-    };
-
-    let mut all_properties = user_properties;
-    all_properties.extend(metadata_properties);
-
-    retain_caller_visible_tags(&mut all_properties, &user_context.user_id);
-
-    let response = EntityPropertiesResponse {
-        entity_id: entity_id.to_string(),
-        properties: all_properties,
-    };
-
-    tracing::info!(
-        entity_id = %entity_id,
-        properties_count = response.properties.len(),
-        "successfully retrieved entity properties"
-    );
-
-    Ok(Json(response))
-}
