--- rust/cloud-storage/properties_service/src/api/properties/entities/get_bulk.rs
@@ -1,182 +1,0 @@
-use axum::{
-    Extension, Json,
-    extract::State,
-    http::StatusCode,
-    response::{IntoResponse, Response},
-};
-use std::collections::HashMap;
-use thiserror::Error;
-
-use crate::api::{
-    context::PropertiesHandlerState,
-    properties::entities::types::{BulkEntityPropertiesRequest, EntityPropertiesResponse},
-    properties::properties_err_status,
-};
-use model::user::UserContext;
-use properties::{PropertiesErr, PropertiesService};
-
-#[derive(Debug, Error)]
-pub enum GetBulkEntityPropertiesErr {
-    #[error(transparent)]
-    Properties(#[from] PropertiesErr),
-    #[error("Entities array cannot be empty")]
-    InvalidRequest,
-    #[error("Access denied")]
-    Permission(#[from] crate::api::permissions::PermissionError),
-}
-
-impl IntoResponse for GetBulkEntityPropertiesErr {
-    fn into_response(self) -> Response {
-        let status_code = match &self {
-            GetBulkEntityPropertiesErr::Properties(e) => properties_err_status(e),
-            GetBulkEntityPropertiesErr::InvalidRequest => StatusCode::BAD_REQUEST,
-            GetBulkEntityPropertiesErr::Permission(e) => e.status_code(),
-        };
-
-        if status_code.is_server_error() {
-            tracing::error!(
-                error = ?self,
-                error_type = "GetBulkEntityPropertiesErr",
-                "Internal server error"
-            );
-        }
-
-        (status_code, self.to_string()).into_response()
-    }
-}
-
-/// Shared implementation for bulk entity properties retrieval
-async fn get_bulk_entity_properties_impl(
-    state: &PropertiesHandlerState,
-    request: BulkEntityPropertiesRequest,
-) -> Result<HashMap<String, EntityPropertiesResponse>, GetBulkEntityPropertiesErr> {
-    if request.entities.is_empty() {
-        tracing::error!("empty entities array in request");
-        return Err(GetBulkEntityPropertiesErr::InvalidRequest);
-    }
-
-    tracing::info!("retrieving bulk entity properties");
-
-    // An empty property_ids fetches all properties for the given entities.
-    // Note: the public endpoint requires property_ids, but internal callers can
-    // pass an empty vec to fetch all properties for the given entities.
-    let bulk_properties = state
-        .properties_service
-        .get_bulk_entity_properties(request.entities, request.property_ids)
-        .await?;
-
-    let mut result: HashMap<String, EntityPropertiesResponse> = HashMap::new();
-
-    for (key, properties_values) in bulk_properties {
-        result.insert(
-            key.entity_id.clone(),
-            EntityPropertiesResponse {
-                entity_id: key.entity_id,
-                properties: properties_values,
-            },
-        );
-    }
-
-    tracing::info!(
-        successful_entities = result.len(),
-        "successfully retrieved bulk entity properties"
-    );
-
-    Ok(result)
-}
-
-/// Get properties for multiple entities in bulk (internal endpoint - service-to-service)
-#[utoipa::path(
-    post,
-    path = "/internal/properties/entities/bulk",
-    request_body = BulkEntityPropertiesRequest,
-    responses(
-        (status = 200, description = "Bulk entity properties retrieved successfully", body = HashMap<String, EntityPropertiesResponse>),
-        (status = 400, description = "Invalid request body"),
-        (status = 500, description = "Internal server error")
-    ),
-    tag = "Internal"
-)]
-#[tracing::instrument(skip(state, request), fields(entity_count = request.entities.len()), err)]
-pub async fn get_bulk_entity_properties_internal(
-    State(state): State<PropertiesHandlerState>,
-    Json(request): Json<BulkEntityPropertiesRequest>,
-) -> Result<Json<HashMap<String, EntityPropertiesResponse>>, GetBulkEntityPropertiesErr> {
-    get_bulk_entity_properties_impl(&state, request)
-        .await
-        .map(Json)
-}
-
-/// Get properties for multiple entities in bulk (public endpoint with user auth)
-///
-/// Only returns properties for entities the user has view permission for.
-/// Entities without permission are silently omitted from the response.
-#[utoipa::path(
-    post,
-    path = "/properties/entities/bulk",
-    request_body = BulkEntityPropertiesRequest,
-    responses(
-        (status = 200, description = "Bulk entity properties retrieved successfully", body = HashMap<String, EntityPropertiesResponse>),
-        (status = 400, description = "Invalid request body"),
-        (status = 403, description = "Forbidden - User does not have permission"),
-        (status = 500, description = "Internal server error")
-    ),
-    tag = "Properties"
-)]
-#[tracing::instrument(skip(state, request, user_context), fields(user_id = %user_context.user_id, entity_count = request.entities.len()), err)]
-pub async fn get_bulk_entity_properties(
-    State(state): State<PropertiesHandlerState>,
-    Extension(user_context): Extension<UserContext>,
-    Json(request): Json<BulkEntityPropertiesRequest>,
-) -> Result<Json<HashMap<String, EntityPropertiesResponse>>, GetBulkEntityPropertiesErr> {
-    // Unlike the internal endpoint, the public endpoint requires explicit property IDs.
-    // An empty property_ids means "no properties requested", so return early with empty result.
-    if request.entities.is_empty() || request.property_ids.is_empty() {
-        return Ok(Json(HashMap::new()));
-    }
-
-    // Filter to only entities the user has permission to view
-    let mut permitted_entities = Vec::with_capacity(request.entities.len());
-    for entity_ref in &request.entities {
-        match crate::api::permissions::check_entity_view_permission(
-            &state,
-            &user_context.user_id,
-            entity_ref,
-        )
-        .await
-        {
-            Ok(()) => permitted_entities.push(entity_ref.clone()),
-            Err(e) => {
-                tracing::debug!(
-                    entity_id = %entity_ref.entity_id,
-                    entity_type = ?entity_ref.entity_type,
-                    error = ?e,
-                    "user lacks permission, skipping entity"
-                );
-            }
-        }
-    }
-
-    tracing::info!(
-        permitted = permitted_entities.len(),
-        "filtered entities by permission"
-    );
-
-    if permitted_entities.is_empty() {
-        return Ok(Json(HashMap::new()));
-    }
-
-    let filtered_request = BulkEntityPropertiesRequest {
-        entities: permitted_entities,
-        property_ids: request.property_ids.clone(),
-    };
-
-    let mut result = get_bulk_entity_properties_impl(&state, filtered_request).await?;
-    for response in result.values_mut() {
-        crate::api::properties::entities::types::retain_caller_visible_tags(
-            &mut response.properties,
-            &user_context.user_id,
-        );
-    }
-    Ok(Json(result))
-}
