--- rust/cloud-storage/properties_service/src/api/properties/entities/mod.rs
@@ -1,11 +1,0 @@
-pub mod delete_entity;
-pub mod delete_property;
-pub mod get;
-pub mod get_bulk;
-pub mod option;
-pub mod set;
-pub mod set_property_status_complete;
-pub mod types;
-
-// Re-export commonly used types
-pub use types::{EntityPropertiesResponse, EntityQueryParams, SetEntityPropertyRequest};
