--- rust/cloud-storage/properties_service/src/api/properties/options/mod.rs
@@ -1,4 +1,0 @@
-pub mod create;
-pub mod delete;
-pub mod get;
-pub mod update;
