+++ rust/cloud-storage/properties_service/src/lib.rs
@@ -1,4 +1,8 @@
 //! Properties Service Library - exposes API routes for integration into other services
+//!
+//! The routes themselves live in `properties::inbound::axum_router`; this crate
+//! is the composition shim binding them to the concrete service types and the
+//! service's authentication middleware.
 
 pub mod api;
 
@@ -6,5 +10,5 @@ pub mod api;
 pub use api::context::EntityAccessServiceType;
 pub use api::context::PropertiesHandlerState;
 pub use api::context::PropertiesService;
-pub use api::properties::router as properties_router;
+pub use api::router as properties_router;
 pub use api::swagger::ApiDoc as PropertiesApiDoc;
